+++ webpack-es6.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    mode: "production",
+    entry: "./index.js",
+    output: {
+      library: 'csbiginteger',
+      libraryTarget: 'var',
+      filename: './csbiginteger.mjs',
+      auxiliaryComment: 'csBigInteger.js library'
+    }
+}